♻️  zb: Split uncached property test into its own mod
